added middleware to routes, cleaned controllers, profile -> my-activity, some controller fixes, translations

+++ routes/web.php
@@ -30,21 +30,21 @@ Route::middleware(['setLocaleFromSession'])->group(function () {
     Route::get('/artists', [ArtistController::class, 'index'])->name('artists.index');
     Route::get('/artists/{artist}', [ArtistController::class, 'show'])->name('artists.show');
 
-    Route::get('/albums/create', [AlbumController::class, 'create'])->name('albums.create');
-    Route::post('/albums', [AlbumController::class, 'store'])->name('albums.store');
+    Route::get('/albums/create', [AlbumController::class, 'create'])->middleware('auth')->name('albums.create');
+    Route::post('/albums', [AlbumController::class, 'store'])->middleware('auth')->name('albums.store');
     
-    Route::get('/my-albums', [ArtistAlbumController::class, 'index'])->name('artist.albums.index');
-    Route::get('/my-albums/{album}/edit', [ArtistAlbumController::class, 'edit'])->name('artist.albums.edit');
-    Route::put('/my-albums/{album}', [ArtistAlbumController::class, 'update'])->name('artist.albums.update');
-    Route::delete('/my-albums/{album}', [ArtistAlbumController::class, 'destroy'])->name('artist.albums.destroy');
+    Route::get('/my-albums', [ArtistAlbumController::class, 'index'])->middleware('auth')->name('artist.albums.index');
+    Route::get('/my-albums/{album}/edit', [ArtistAlbumController::class, 'edit'])->middleware('auth')->name('artist.albums.edit');
+    Route::put('/my-albums/{album}', [ArtistAlbumController::class, 'update'])->middleware('auth')->name('artist.albums.update');
+    Route::delete('/my-albums/{album}', [ArtistAlbumController::class, 'destroy'])->middleware('auth')->name('artist.albums.destroy');
 
     Route::get('/albums/{album}', [AlbumController::class, 'show'])->name('albums.show');
 
-    Route::post('/albums/{album}/reviews', [ReviewController::class, 'store'])->name('reviews.store');
-    Route::post('/albums/{album}/ratings', [RatingController::class, 'store'])->name('ratings.store');
-    Route::put('/albums/{album}/ratings/{rating}', [RatingController::class, 'update'])->name('ratings.update');
+    Route::post('/albums/{album}/reviews', [ReviewController::class, 'store'])->middleware('auth')->name('reviews.store');
+    Route::post('/albums/{album}/ratings', [RatingController::class, 'store'])->middleware('auth')->name('ratings.store');
+    Route::put('/albums/{album}/ratings/{rating}', [RatingController::class, 'update'])->middleware('auth')->name('ratings.update');
 
-    Route::get('/profile', [UserController::class, 'index'])->middleware('auth')->name('users.index');
+    Route::get('/my-activity', [UserController::class, 'index'])->middleware('auth')->name('users.index');
 
     Route::get('/admin', [AdminController::class, 'index'])->name('admin.index');
     Route::delete('/admin/reviews/{review}', [AdminController::class, 'destroyReview'])->name('admin.reviews.destroy');
